Pull visualizer Extension out of BaseVisualizer.
